Make `scripts/*` pass ShellCheck
Warning SC2155 is that `local foo=$(cmd)` shadows the return code from
the command, but we weren't checking this anyway, so it's ignored.

Deprecated uses of `cmd` were switched to the updated `$(cmd)`.

Instances `for i in $(cat file)` were replaced with `while read`
loops.

Variables were quoted as needed.

Uses of POSIX-style `[` were replaced with `[[`.

Use of `<<` was replaced with `<` when we didn't mean to append.

Check of `if [[ $? -ne 0 ]]` were replaced with `if ! command;`.

Dangerous `cd` was guarded with `cd <dir> || exit 1`.

Use of deprecated `egrep` was replaced with `grep -E`.

Spacing was fixed up.