Fix format-code to check version correctly
There was a bug in the version check logic that would unconditionally
compare all values; so 3.6.0 was seen as greater than 5.0.0, since
although the major components compared as 3 < 6, it continued on to
erroneously compare the minor components as 6 < 0 and so failed.
Instead, the logic should be that the loop only continues to check
minor and patch if the prior component was equal, not less than.

This resolves #710.