Fix `format-code` to explicitly require `clang-format-3.8`
Because `clang-format` is not invariant between versions, we must
require the exact version our CI is using, which is 3.8 (the default
on Ubuntu 16.04 with no extra repos added). To make things a bit
easier, the prereqs script now explicitly installs `clang-format-3.8`
which makes a binary available by the same name, and the `format-code`
script tries this binary first.

Unfortunately, adding the LLVM 7 snapshot repo (as our prereqs script
does) changes this behavior such that the `clang-format` package
installs the unreleased 8.x clang-format. On Ubuntu 18.04, it is
difficult to get 3.8; you have to extract it from the old release
tarball. However, we will have to address this later (perhaps updating
to `clang-format-6.0`, which is available on both 16.04 and 18.04).

This resolves #711 (pending a better solution).