Add .gitattributes file to handle line-endings
Text files will use automatic line-endings (LFs in the repo, LFs when
checked out on Linux, CRLFs when checked out on Windows).

Shell scripts explicitly will always checkout with LFs, even on Windows,
so that WSL can execute them even when Git for Windows was used, or if
the files are cloned on Windows and then copied to Linux.

Windows build files will always have CRLFs so that if the repo is cloned
on Linux, and then the files are mounted to a share read on Windows, the
files will have the expected line endings for the Windows tools (reading
them remotely, or copying them).