Change sentence to be in line with the explanation
The first sentence was a bit "over-promising" as it is hard to define
what "the whole community agrees" really means. The later explanation
goes on to explain that the intent is the community provides feedback,
and it will all at least be discussed, but this should be reflected in
our first summary sentence too.