Add a "defaults" stage to CI
This stage tests the code as if a developer was "getting started." That
is, it uses our `install-prereqs` and `check-precommit-reqs` scripts,
and does not set any environment variables which would affect the
compilation.

Note that it builds a Docker image on the fly with the given minimal
Dockerfile because Jenkins runs Docker with a mapped user, not as root,
so package installation must happen before Jenkins enters the container
for the build.