OE_TRACE_LEVEL cmake parameter.
1. OE_TRACE_LEVEL value can be specified to cmake and it is propagated
to compiler definitions.

2. If OE_TRACE_LEVEL is info, any OE_CHECK failure is logged.
This allows quickly identifying the root cause of an attestation failure,
and other kinds of failures where the OE_CHECK output may not be readily
available/printed at the top-level.

3. If cert verification failed, with tracing enabled, print the error message
reported by the crypto library.

4. Make some tests robust to the OE_TRACE_LEVEL specification by undef-ing incoming define.

Note OE_TRACE_LEVEL=0 is equivalent to not defining it. Use OE_TRACE_LEVEL=2 for maximum information.
If tracing is enables, few tests that rely on the printed out strings fail. This is expected since
tracing changes the printout.