Remove accidentally added `C` and `CXX` compile definitions
This target had `C` and `CXX` added as compile definitions on accident
because the programmer was using `target_compile_definitions` as if it
were `add_compile_defs`. The CMake built-in `target_compile_definitions`
does not accept languages as parameter, and so this logic instead was
adding `C` and `CXX` as definitions.