Fix missing dependency on oeedger8r.
Use temporary workaround to ensure that if edger8r source changes,
build tests that depend on it.
The oeedger8r target depends on the edger8r binary, tests depend on
the orrdger8r target.
When building a test, the oeedger8r target is built if
the edger8r source has changed.
However the newly built edger8r is not run on the test's edl file.
This is because the oeedger8r target does no (and has no way) to declare
that its output is the edger8r binary.

This changelist incorporates a workaround to fix the above issue.
Tests also depend on the oeedger8r binary explicitly.