Fix tests/libcxx test exclusion
PR #889 reintroduced 2 libcxx tests that fail on Clang release builds,
with changes to CMakeLists.txt to exclude them from those configurations.
The check for build type was not case insensitive and so fails to exclude
the tests during CI tests. This PR fixes that by forcing upper case
comparisons of the build type.