Allow enclave executables to use both C and C++.
Previously, oelibcxx unconditionally added libcxx to the include path.
When a C file includes a header that is common between C and C++ say
stdint.h, the C++ version would be picked up which has C++ syntax
(extern "C") and causes compile errors.

This fix changes 3rdparty/libcxx so that the c++ include path is added
only if the file being compiled is a cpp file. This is enforced
via the COMPILE_LANGUAGE:CXX generator expression.