Set CI timeouts
The Bors timeout has been upped to 2 hours because it includes the time
for a build to start, which with our bogged CI system can be a while.

The Jenkins timeouts have been set by analyzing the `buildTimeTrend`
report for each build, such that if the timeout is hit, the build was
anomalous. For most builds that means 30 minutes, with the exception of
the slow full C++ tests which is 3 hours. Note that occasionally that
build will "successfully" complete in 4-6 hours, but this should be
considered a failure because it means something hung for a while.