Move CI timeouts to each stage
Instead of applying to the pipeline, which could cause a failure due to
a long queue as the stages run in parallel (and start as executors
become available), we instead set the timeout for each stage
individually, such that it only counts the build time. This also allows
us to fine-tune the timeout for each stage.