Dynamically configured testing of basic types.
The types wchar_t, long and lond double have different sizes
on Linux and Windows. The test is now refactored so that
testing of these types is enabled on the fly, if both the
host and the enclave have the same sizes for these types.

In the future, all the non-sized  basic types (char, int etc)
can be tested this way. This will be applicable if say the host is
32 bit, enclave is 64 bit etc.