Move pre-commit requirements stage into parallel build
It was determined that a previous CI check erroneously failed because it
was restarted in-between stages. The Jenkins log stated:

> Stage "Precommit-checking" skipped due to this build restarting at
> stage "Build and Test"

Now, I still don't know why a Jenkins Pipeline is allowed to succeed if
a stage is skipped, but we can at least prevent this stage from being
skipped by adding it to the parallel stage.

Note that this also enables fail-fast semantics, so that the build fails
as soon as any one stage running in parallel fails. This returns faster
results to users, and frees up resources.