Fix translation aborts on ARMv7 and ARMv8 when -Os is specified. (#53)
TA's are compiled by specifying the relevant source files and including the TA dev kit
as generated by OP-TEE's build system. The dev kit may differ from platform to platform
and may or not specify that TA's be built with optimizations. The OE_SET_OUT_POINTER
macro sets an element of the input arguments marshalling structure by casting the LHS to
a pointer to a pointer to a byte array and sets it to a different value. This confuses
GCC which in turn forgoes updating the register whose value is that of the element in
the structure before invoking the ECALL's implementation proper. This means that the
structure's value in memory is correct eventually but the value passed to the ECALL is
garbage, which causes the translation aborts.