Correctly scope -Wno-error=unused-command-line-argument
This commit moves the assembly sources out of `oelibc` and into a CMake
"object" library, `oelibasm`. This allows us to set the flags for the
assembly sources separately. While this removes the `-isystem` and some
other flags from the generated assembler command, we are still stuck
with a bug due to CMake's default inclusion of `-DNDEBUG` etc., which
clang-7 warns about when used as an assembler. Since we treat warnings
as errors, and can't fix this warning, we just have to ignore it.

Previously we ignored the error for everything. Now we just ignore it
for the assembly files.