Create _v2 of APIs that have bad calling style
Issue https://github.com/Microsoft/openenclave/issues/841 states that we
should return buffers directly instead of passing in a buffer. The old
style overloads the APIs to first get the length and then call in with a
buffer. The V2 versions of the APIs allocate the memory and returns
that, adding a function to free the returned buffers.