oe-gdb: Support debugging in simulation mode.
In simulation mode, we were creating write-protected pages to load
the enclave. This made it impossible for gdb to set a breakpoint.
GDB tries to insert an 'int 3' instruction in enclave code but since
the code resided in write-protected memory, the insert fails.
To support debugging, in simulation mode, the enclave pages
are made writable using the PROT_WRITE flag.
The oe-gdb test is also always run in simulation mode as well.