Make get/parse/verify_report host APIs use oeedger8r (#1141)
Previously they used sgx_edger8r

This also creates a way for other internal APIs to use oeedger8r
without requiring that the application's EDL include an internal
EDL file.  However, to do this we have to work around a current
limitation of oeedger8r where it generates a common ECALL table,
where we need two separate ECALL tables.  This change includes
such a workaround (redefining the names of the ECALL tables in
an include file that is included from generated code).  This
workaround can go away if oeedger8r is updated with some way to
generate a separate ECALL table.

In the future, more APIs in oeoverintelsgx.edl can move to this
mechanism, which will require a similar workaround for OCALLs
at that time.

Signed-off-by: Dave Thaler <dthaler@microsoft.com>