Fix mbed TLS patch step
This was a bit confusing as a few things had to be changed to get the
dependencies correct. First, we should be using `configure_file` to
actually configure the file (with `#cmakedefine`s in it) instead of just
copying the file, hence the `file(APPEND)` goes away. Second, we need to
specify an `UPDATE_COMMAND` for the external project so that it
configures and builds again if its sources (including the configured
file) change. This actually replaced the `PATCH_COMMAND`. Third, when
the sources of the external project change, while we want it to
configure and build again, we do not want it to "install" (or copy) the
headers again, as this triggers an unnecessary rebuild of all targets
that use the headers. Instead, the headers and built libraries are
referenced directly out of the source tree of the external project, and
its install step is turned into a no-op. This causes dependent libraries
to link again if the mbed TLS libraries are changed, but not to build
again (since the headers did not change). This actually simplified the
header logic such that the copy now only happens during installation.