Clean up Doxygen CMake rules
This uses the modern version of the Doxygen CMake package, which
replaces the use of `ExternalProject` and moves the configuration from a
machine-generated file to the `CMakeLists.txt` file.

A slight change to the output of the docs was made: the base directory
no longer uses `include/openenclave` as this is common to tall the
docs (and is implied).

The `refman` target was renamed to `doxygen`, and the script was
updated. A slight change was made to the script to remove some redundant
code, and fix the output directory.

squash! Clean up Doxygen CMake rules