Thread-Local Storage in Simulation mode.
In simulation mode, carefully set the GS and FS
registers upon entry/exit from an enclave.
These registers are set in SGX mode by EENTER and EEXIT
instructions, but in simulation mode, the host has
to do it carefully.
The FS register is key for libc, pthreads etc to work.
And therefore, FS register must be restored as early
as possible upon re-entry to the host.

Also fixed a bug where if an OCALL made an ECALL to
*another* enclave, the thread-binding was not being
restored correctly.