Build all supported tests on Windows using clang cross-compilation
1. Use BUILD_ENCLAVES instead of UNIX to determine if test enclaves
need to be built or not.
2. MUSL disable -Wconversion warning in endian.h. This is needed because
unlike linux, in windows the cmake generator does not add the -isystem
prefix for the libc includes which would suppress the warning by treating
endian.h as a system header.
3. add_enclave.cmake calls maybe_build_using_clangw if USE_CLANGW is
set. Additionally it propagates the cmake variables set by
maybe_build_using_clangw back to the caller.
4. ld.lld aggressively optimizes initial-exec threading-model to local-exec
threading-model. Hence the assertion for relocation entries is disabled in
host in this workflow.