Rewrite check-license script to be fast
This script now works instantaneously when given a small list of files.
When working across the entire repository, it takes just a few seconds.

Fixing this required changing the ignore logic to use regular
expressions instead of shell globs so that we could perform the ignore
checks much more quickly.

The script now also asserts that the license is found in the first four
lines of the file, which was a bug in the previous implementation.