Fix format-code script to accept list of files
This fixes the script to accept a space separated list of files (like
the other arguments, not a comma), and checks the user provided list
against the specified file extensions. This is necessary so we can pass
a list of modified files and have the script ignore those that
irrelevant, which makes this usable in a Git pre-commit hook.