Fix for correctly throwing std::bad_alloc
LLVM's exception.cc uses calloc to create the thread_info structure
to maintain the state. If the enclave has run out of memory and the
first exception being thrown is std::bad_alloc, then the calloc to
create thread_info would itself fail resulting in a SEGV when the
thread_info is later used. The fix is to use a predefined block
of memory reserved in the oe_thread_data_t structure.
Note: Intel SDK uses a similar strategy.

Locked down std::bad_alloc with a test.

Added README.md for catalog modifications to libcxxrt.