Add INSTALL_PREFIX to INSTALL_INTERFACE variables
This fixes the export file to correctly have `${_IMPORT_PREFIX}` as the
prefix for otherwise relative paths. Doing this requires using the CMake
generator expression `$<INSTALL_PREFIX>` which is replaced by the former
variable in the export file. It still needs to be followed by the
`GNUInstallDirs` variables e.g. `${CMAKE_INSTALL_LIBDIR}` which
otherwise is replaced by e.g. `lib`.