Fix comparison error
This has to be >=, obviously. It was missed in testing since originally
no version check was being performed, hence it passed locally, and then
a version check was added to make 16.04 CI pass. Unfortunately, we don't
yet have 18.04 GCC CI yet, so it was not caught and I must not have
manually tested it (we've been distracted by a massive CI outage).