Fix enclave/core/__secs_to_tm.c compilation
Unfortunately, `time_impl.h` needs the `time.h` definition from
`src/include/time.h`.

`enclave/core/__secs_to_tm.c` needs a very limited subset of functionality in
the MUSL version, and what has changed in 1.1.21 is that the headers on which
the MUSL implementation depends on, now requires us to reconcile more symbols
that are not actually used but need to be present for the compiler.

Signed-off-by: Alin Gabriel Serdean <aserdean@ovn.org>
Co-authored-by: @CodeMonkeyLeet
---
v3: fixup code formatting
v2: fold in @CodeMonkeyLeet suggestion, add him as co-author and change the
commit message.