Ansible: Use GCC 5 on Ubuntu 1804
The default GCC version from Ubuntu 1804 is GCC 8 and it causes
failures with the full libcxx unit tests suite.

This change makes the `gcc` and `g++` binaries from Ubuntu 18.04
to point to the version 5 binaries from the system.

Thus, we end up with the same GCC compiler used on both Ubuntu
16.04 and 18.04.