Export `oeedger8r` in the CMake package
This is a work-around to export the `oeedger8r` as a target usable by
users of the CMake package. We cannot use the normal `install(TARGETS)`
method to export this target since it, being an OCaml binary built via
custom commands instead of C++ built with `add_executable`, is an
imported target in the upstream (our) build.