Improve oesign arguments parsing
Improve the `oesign` CLI arguments parsing using `getopt_long`.

Instead of relying on the arguments order and name, the
following mandatory flags are added to the `sign` subcommand:
* `--enclave-image [-e]`, the enclave image file path.
* `--config-file [-c]`, the path of the config file with
  enclave properties.
* `--key-file [-k]`, the path of the private key file used to
  digitally sign the enclave image.

The `dump` subcommand has been updated as well, and it accepts
only the `--enclave-image [-e]` mandatory flag.

The samples and documentation have been updated for the latest
`oesign` changes.

Co-authored-by: Alin Gabriel Serdean <aserdean@ovn.org>
Signed-off-by: Ionut Balutoiu <ibalutoiu@cloudbasesolutions.com>