Fix compiler warning in oesign
When `-fno-builtin` was removed from oesign by nature of it being scope
to just enclave code in our build, this (correct) compiler warning
showed back up. There is a funny compiler interaction where this warning
was suppressed when builtins were disabled, but it is a valid warning so
we should just fix it.