Fix type punning errors in generated code
When building with strict-aliasing enforced, our generated code produced
the warning/error "dereferencing type-punned pointer will break
strict-aliasing rules" because we assumed all pointer types were
compatible with `uint8_t`. While this might be a safe assumption, the
compiler disagrees, and there is no reason for us to make it. Instead,
we just pass the type to the macros when generating the code so that our
casting is correct.