[CI/CD] Fix CI shared lib race condition
Variables declared in Jenkins groovy pipeline logic without any
type or the keyword `def` are considered global by default.
Every global variable is shared among parallel stages with all
the nodes.

The `image` variable from the CI shared library was meant to be
a local variable, but due to the fact that we lacked `def` or
any type when declaring the variable, this was considered global
and caused a lot of race conditions, and random transient failures.