Reuse `get_cast_to_mem_expr` for casting marshalling buffers
After further research, it turned out this existing function did exactly
what I was looking for when coming up with a fix. We need to take a
parameter type and cast it to the member expression for the member in
the marshalling struct. This is a weird cast because arrays must be cast
to their root type. That is, my prior TODO of casting `int[][]` to
`int**` would have been incorrect, as the actual member in the
marshalling struct would be `int*`.