Refactor `get_cast_to_mem_expr`
This combines it with the similar (copied, but modified) function below
by adding an argument to control whether or not the surrounding
parentheses are added.

This also fixes spacing of cast expressions. That is, our style does not
have a space between the cast expression and the item being casted.