Isolate all I/O side effects into one location
This changes `open_file` to `write_file` which is the only function that
performs file I/O. Then, all the 'gen' functions simply return their
list of strings (becoming variables instead of functions), and it is
only at the end of `gen_enclave_code` that conditional file I/O is
performed.

This has the side effect (heh) of the string lists moving up an
indentation level. This change was required in order to remove the
arguments to the 'gen' functions, because when they turned into
variables, the were unconditionally evaluated, and the embedded side
effects became a problem.