Add regression tests for `count * size` marshalling
This adds two tests checking that we don't over-copy the number of bytes
requested when the `count * size` feature is used. This is a weird edge
case where we allow the specification of the total bytes to copy as
being the product of `count` and `size` if they are both specified.
These tests check the regression that we were defaulting to copying
`count` members instead of `count * size` bytes, since we weren't
explicitly checking if both `count` and `size` were specified.