Handle `count * size` behavior
Now we explicitly check for and handle the case where both `count` and
`size` attributes were specified, emitting their product for `size`, and
their product divided by the size of the member for `count`. We should
add checks if the size is not an even multiple of the member size.