[Ansible] Downgrade Git version to 2.19.1
We use extra tools bundled with the Git Windows installation.

Git 2.20.1 comes with OpenSSL version 1.1.1, and it caused
us the regression https://github.com/microsoft/openenclave/pull/1932 after the Windows CI servers were upgraded
to this version.

Git 2.19.1 comes with OpenSSL version 1.0.2, and it doesn't
reproduce the above regression. Therefore, until a proper fix is
submitted, we want to use this version of Git Windows installation.