Add Visual Studio 2019 extension
This allows supporting the new ELF debugger if used with the 2019
preview.

A separate nupkg (open-enclave-cross) is used that includes the VS
targets files needed by this extension.

Includes support for compiling static libs and C++ files in enclaves

clang is used to compile enclaves and Linux host apps, while the
VS compiler is used for Windows host apps

x64 is supported, but x86 is not

The templates use OE_SET_ENCLAVE_{SGX,OPTEE} macros, but also
the .conf files are kept so that the developer can see how they can be
optionally used, via comments in those files.

The strong name key (key.snk) is checked in per recommendation at
https://docs.microsoft.com/en-us/dotnet/standard/assembly/strong-named?redirectedfrom=MSDN
and
https://stackoverflow.com/questions/36141302/why-is-it-recommended-to-include-the-private-key-used-for-assembly-signing-in-op

Signed-off-by: Dave Thaler <dthaler@microsoft.com>