OE_AEP_ADDRESS variable to obtain precise location of ENCLU.
For single-step debugging to work, OE_AEP must exacly point to the
ENCLU instruction.
If used as a function with type void (void), the linker could emit
thunks to wrap the OE_AEP function (say for incremental linking).
To obtain the precise location of the ENCLU, the OE_AEP_ADDRESS
variable is used to store the computed address of the ENCLU instruction.

Note: OE_AEP is not a proper C function. It does not confirm to the
Linux/Windows ABI.