DebugRt: Stack of thread bindings (oe_debug_thread_bindings_list)
Everytime an ecall is made, a new triplet is created
   (current-thread-id, debug-enclave*, tcs)
and pushed into a stack maintained by the debug runtime.

When the ecall returns, the top most binding for the current
thread is popped.

A debugger can scan this global stack (actually a list)
of bindings an can figure out the tcs-es for the enclave in/out
transitions in all the threads.

Note: In future, we may need to push ecall/ocall landing frames
rather than just the triplet.