Resolve some old TODOs
The note about checking if `hd[i] == NULL` is actually wrong. It is
sufficient to check if `hd == NULL`, and then any member arrays under
that. Unfortunately, there's nothing to do about checking if the user
supplied bounds match up to the user supplied data, but that's C for
you.