Rename Community Maintenance Committee
Given the confusion between Maintainer vs CMC member, Maintainers Group
for example might be clearer.  Mike and John indicated initial support
for this term.

Another term used by many projects is "Technical Steering Committee",
which is the term that also appears in Stephen Walli's doc.   I would
be ok with that term, just for consistency with other open source
projects, if the group prefers that term.  However, it has the same
confusion between Maintainer vs TSC member.   One way that could be
resolved, should the TSC name be used, would be to remove the term
"Maintainer" and only use "TSC member", "Committer", and "Contributor"
as roles, which is the approach Stephen's doc takes.

Signed-off-by: Dave Thaler <dthaler@microsoft.com>