Plumb pointer type through `oe_gen_set_pointers`
We were previously assuming that all nested pointers were "in/out" types
as they did not contain the information necessary to determine if they
were just "in" or "out" types. This information exists only at the
top-level definition of the function (the pointer parameter is either
"in," or "out," or "in/out"). So we can instead determine the correct
setter before the first recursion, and then continue to pass it through.

This is made easier by the introduction of `flatten_map2`, which acts
just like `flatten_map` but for `List.map2`.

Signed-off-by: Andrew Schwartzmeyer <andrew@schwartzmeyer.com>