Add sleep to tls_server_enc signing to allow header to be generated in `attested_tls` sample.
The sleep is added because cmake opens bash in a separate process on
Windows. This means that the build dependecy on the autogenerated
tls_server_enc_mrenclave.h file may be satisfied and files which depend
in it may be compiled before the file is actually generated. This causes
transient build failures for the attested_tls sample.

Also adds tls_server_sign_enc as a dependency for both tls_client_enc
and tls_non_enc_client. This ensures that those targets will not be
built before the target that generates tls_server_enc_mrenclave.h has
been run.

Signed-off-by: Jordan Hand <jorhand@microsoft.com>