Exposed backtrace via GNU functions (execinfo.h)
Compile with -fno-omit-frame-pointer. This allows using rbp
to walk up the stack.

Backtrace functionality is exposed via the GNU functions instead of
oe_ prefixed funtions.

A tricky aspect of the implementation is that backtraces are used
by debug malloc (which is available only in debug builds).
When debug malloc uses backtraces, it must not use (debug) malloc
to allocate memory; instead it must use the lower level dlmalloc.
This is enabled by passing memory management functions to
oe_back_trace_symbols_impl.

To avoid duplication of the backtrace gathering logic, the implementation
is moved into oe_backtrace_impl. It is expected to be called from
oe_backtrace and from backtrace. It removed the frame of the caller
in the gathered backtrace.

Signed-off-by: Anand Krishnamoorthi <anakrish@microsoft.com>