Implement global module initialization functions for OE modules
OE modules are static libraries that provide discreet functionality (such
as libc or a host filesystem) which are compiled as static libraries and
linked into the enclave binary. Modules may have any number of
initialization functions, which will be called on first enclave entry
prior to global constructors.

Implements module initialization and documents this feature.

Signed-off-by: Jordan Hand <jorhand@microsoft.com>