Allow libc tests to be run individually
The libc tests currently run all as a single test and if one fails then
the tests stop running. Add the cmake option OE_SEPARATE_LIBC_TESTS
which will lead ctest to run each libc test as a separate case.

Additionally, add a command-line option to the libc test host to allow a
single test to be run at a time to support the cmake flag.

Signed-off-by: Jordan Hand <jorhand@microsoft.com>