Add @advanced doxygen tag support
Similar to the support for @experimental added in PR #2364,
this PR adds support for @advanced in doxygen to help address
issue #2547.

This change causes a separate "advanced.html" page to be created that has
a list of all advanced APIs. Tagging an API with @advanced then causes
the following notice to show up on that API's own page, with the notice
hotlinked to the advanced.html page:

```
This feature is marked as advanced and is not intended for use by apps:
```

Signed-off-by: Dave Thaler <dthaler@microsoft.com>