Debugger Contract: SDK performs Ocall Stack Stitching
Confirming to the Debugger Contract, the SDK performs ocall stack stitching.
This simplifies various debugger implementation.

As part of this PR, the Linux debugger (oegdb) has been made to confirm to the
debugger contract. And thus the python code to perform the stack stitching has
been removed. As a result, the debugger no longer relies on any OE specific
data structures in the enclave side.

In order to allow the host to perform stack stitching, the enclave passes its
eexit frame information to the host via the host's ecall_context for the thread.
An enclave can securely determine if it is debuggable (trace/logging already does this).

The host's ocall dispatching mechanism temporarily stitches the ocall stack prior to
dispatching the ocall and then reverts the stack once the ocall completes.
Previously a similar, but less robust scheme of modifying the call stack was used
by the debugger.

Description of changes:

- asmdefs.h
  Declared oe_asm_exit.

- exit.S
  Removed calls to oe_notify_nested_exit_start.
  Removed implementation of oe_exit_enclave (__morestack).
  oe_exit_enclave is implemented in C.

- gdb_sgx_plugin.py
  Removed debugger code to stitch the stack.
  The SDK stitches the stack instead.

- enter.S
  Remove unused code. Ocalls never used this code and the code was dead.

- core/sgx/calls.c:
  Implement oe_asm_exit (__morestack) and __debug_write_eexit_frame.
  __debug_eexit_frame records its caller (__morestack) information and
  passes it along to the host so that the host can stitch the ocall stack.

- host/sgx/enter.c
  __oe_host_stack_bridge stitches the ocall stack before dispatching the ocall function.
  After the function completes, the call stack is restored.

- ecall_context.h
 Add fields for eexit frame information for use in ocall stack stitching.

- tests/debugger/oegdb
  Added test to lockdown ecall and ocall stack stitching.
  An ocall function which is called by an ecall function which is in turn
  called by main, all 3 have volatile local magic variables initialized to
  zero. The debugger is expected to break in the ocall, walk the ocall and
  ecall stacks and set these magic variables to MAGIC_VALUE. If the debugger
  fails to walk the stack correctly, the magic variables will retain their
  original value (0) instead of the expected value (MAGIC_VALUE), the test
  would fail.

  Also removed enclave side contract assertions.
  The debugger depends only on SGX data structures on the enclave side.

Signed-off-by: Anand Krishnamoorthi <anakrish@microsoft.com>