Link all host code with noexecstack
Compile all host code on linux with noexecstack to ensure the stack is
not executable.

Not only is this good security practice as the Linux kernel invokes
READ_IMPLIES_EXEC on mmap and mprotect for processes with executable
stacks, which means read requests will be treated as read/execute
requests. The SGX driver however ignores the READ_IMPLIES_EXEC
flag when hooking mmap/mprotect calls which can causes spurious failures
when attempting to mmap/mprotect enclave pages.

Signed-off-by: Jordan Hand <jorhand@microsoft.com>