Simulation Mode in Windows
This is a long standing feature request.
Simulation mode reduces the barrier to start playing with the OE SDK.

We had held off on implementing the simulation mode earlier because
- Simulation mode requires ability to manipulate FS and GS registers.
  Though it is possible to change these registers, Windows thread-scheduler
  reverts them back in the middle of enclave executation.
- We didn't yet have thread-local storage and other features like stack protectors
  that depend on FS register
- The enter/exit functions were difficult to change.

With @nonpolarity's work to implement stack protectors (#456f3a2), enclave's no longer
use the GS register and solely use FS register, conforming to Linux ABI.
Since Windows uses GS register to handle thread related data, there is no register conflict
between the enclave (FS) and host (GS).
In simulation mode, only FS register needs to be set before calling the enclave.
GS register can be left unchanged.

Windows thread-scheduler, unlike Linux, however restores the FS register to its original value
whenever the thread resumes execution. This can happen in the middle of an ecall and thus
and enclave could end up with the wrong value of FS register in the middle of an ecall.
FS as set by Windows points to memory not accessible to the application and this results in
access violations.
Simulation mode registers a signal handler (executed first) that determines whether an
exception occurent within a simulation mode enclave and whether FS was incorrect at the
time of exception. If so, the exception handler sets FS to the correct value and requests
the program to continue execution.

TODO:
- Add documention in another PR
- Update CI
- Update debuggers to handle simulation mode

Note:
- Debugging works out of the box under VS
- WinDbg has explict guards to not support simulation mode enclave.
  These guards need to be removed and any missing functionality added.
- When Windows restores the FS register, it often results in enclave code
  throwing an access violation. The debugger can be set up to ignore this
  exception the first time it is hit and the debugging experience is smooth
  afterwards.

Signed-off-by: Anand Krishnamoorthi <anakrish@microsoft.com>