Switchless ECALLs implementation
This is for most part an exact mirror implementation of switchless OCALLs.

The host create switchless ecall worker threads. Each worker thread waits
on the host side for on an event that is triggered if the thread was waiting
at the time an ecall message was posted on the worker thread.
Upon waking up, the thread makes an ecall into the enclave to launch the
enclave portion of the worker thread. On the enclave side, the worker
looks at its context to see if there is an ecall scheduled on it. If so,
it processes the ecall. If not, it keeps spinning until a threshold is reached
at which time, it goes back into the host to wait for the event. This is because
there is no efficient way for an enclave to wait for an event; the OS threading
primitives are not available within the enclave and the wait must happen on the
host side.

- update changelog
- Make oe_handle_call_enclave_function non static to be reused by switchless ecalls
- Add switchless ecall worker ecall
- Update switchless sample to show switchless ecalls
- Update switchless tests to lock down switchless ecalls
- Remove oeedger8r negative test for switchless ecalls
- Remove oeedger8r error for transition_using_threads attribute on ecalls
- Added oe_enclave_worker_context_t type
- Refactored worker wait and wake functions on both Windows and Linux
- Update backtrace test with new internal function name
- Fix memory leak in switchless manager destruction (normal and error scenario)
- atomic.h: added oe_atomic_load, oe_atomic_compare_and_swap_32 and oe_yield_cpu

Signed-off-by: Anand Krishnamoorthi <anakrish@microsoft.com>