Refactor syscall ocalls into smaller groups in multiple edl files
Currently all syscalls must be included as a monolith and are kept in a
single EDL file. Refactor these into smaller groupings so they can be
more easily included to provide discrete functionality when important by
an enclave application.

The ocalls are organized by the POSIX header which contains the
function. For example, if nanosleep() is in time.h, then
oe_syscall_nanosleep_ocall is in time.edl.

Signed-off-by: Jordan Hand <jorhand@microsoft.com>