oe-gdb: Automatically continue execution upon cpuid SIGILL GDB command `oegdb-ignore-sigill <instruction>` to ignore sigill for the given instruction and continue execution.
OE SDK's crypto library (mbedtls) uses cpuid instruction to check for
aesni capability. This causes a SIGILL exception that is eventually
handled by the enclave by emulating the cpuid exception.

This SIGILL however causes the debugger to stop. Anyone that debugs
a crypto application would be first greeted with this SIGILL as the
crypto library initialized.

This PR changes oegdb to continue executing without stopping when
SIGILLs are raised due to cpuid.

Signed-off-by: Anand Krishnamoorthi <anakrish@microsoft.com>