Capture Windows image directly into dest RG
Capture the Azure managed image directly into the target
resource group, and remove the redundant step of `az resource move`.

The `az resource move` was not ideal anyway, since it locks
the target resource group during the operation. This messed the
parallel stages to build CI images, and it adds redundant
execution times.

Signed-off-by: Ionut Balutoiu <ibalutoiu@cloudbasesolutions.com>