Dynamically load Intel SGX Enclave Common shared library
This eliminates that the host application had on Intel SGX Enclave Common Library.

This PR would allow the SDK to be built without the library being present in the
build system. Additionally, it will help make simulation mode work on non-sgx systems.

This based on the same approach we do for the DCAP Quote Provider shared library.
(libdcap_quoteprove.so). See host/sgx/linux/sgxquoteproviderloader.c.

We keep copy of sgx_enclave_common.h in 3rdparty/intel. This describes the
contract between OE SDK and the sgx_enclave_common shared library.

The shared library is lazily loaded via a oe_once function.
Necessary functions are looked up and cached.

NTH:
libsgx_enclave_common.so can provide a version information that the OE SDK
can use to determine if the library conforms to the contract at the time
the SDK was built.

Testing:
1) The existing CI tests will automatically lockdown the case where
   the library exists.
2) We need tests for
   - Case when the library is not present
   - Case when the library is present, but does not have all the
     functions.

Signed-off-by: Anand Krishnamoorthi <anakrish@microsoft.com>